requirements.txt
flask~=3.0.1
mysql-connector-python~=8.3.0
PyYAML~=6.0.1
pytz~=2023.3.post1
shortuuid~=1.0.11
tabulate~=0.9.0
marshmallow~=3.20.2